+++ jsonschema-module-jackson/src/main/java/com/github/victools/jsonschema/module/jackson/JacksonOption.java
@@ -47,6 +47,16 @@ public enum JacksonOption {
      */
     FLATTENED_ENUMS_FROM_JSONPROPERTY,
     /**
+     * Use this option to sort an object's properties according to associated
+     * {@link com.fasterxml.jackson.annotation.JsonPropertyOrder JsonPropertyOrder} annotations. Fields and methods without such an annotation are
+     * listed after annotated properties.
+     */
+    RESPECT_JSONPROPERTY_ORDER,
+    /**
+     * Use this option to skip property name changes according to {@link com.fasterxml.jackson.databind.annotation.JsonNaming JsonNaming} annotations.
+     */
+    IGNORE_PROPERTY_NAMING_STRATEGY,
+    /**
      * Use this option to skip the automatic look-up of subtypes according to {@code @JsonSubTypes} annotations.
      */
     SKIP_SUBTYPE_LOOKUP,