+++ jsonschema-module-jackson/src/test/java/com/github/victools/jsonschema/module/jackson/JacksonModuleTest.java
@@ -19,6 +19,8 @@ package com.github.victools.jsonschema.module.jackson;
 import com.fasterxml.jackson.annotation.JsonClassDescription;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyDescription;
+import com.fasterxml.jackson.databind.PropertyNamingStrategy;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import com.github.victools.jsonschema.generator.ConfigFunction;
 import com.github.victools.jsonschema.generator.FieldScope;
 import com.github.victools.jsonschema.generator.MethodScope;
@@ -26,6 +28,8 @@ import com.github.victools.jsonschema.generator.SchemaGeneratorConfigBuilder;
 import com.github.victools.jsonschema.generator.SchemaGeneratorConfigPart;
 import com.github.victools.jsonschema.generator.SchemaGeneratorGeneralConfigPart;
 import com.github.victools.jsonschema.generator.TypeScope;
+import java.util.List;
+import java.util.stream.Collectors;
 import junitparams.JUnitParamsRunner;
 import junitparams.Parameters;
 import org.junit.Assert;
@@ -61,7 +65,7 @@ public class JacksonModuleTest {
     public void testApplyToConfigBuilder() {
         new JacksonModule().applyToConfigBuilder(this.configBuilder);
 
-        this.verifyCommonConfigurations();
+        this.verifyCommonConfigurations(true);
 
         Mockito.verify(this.configBuilder).forMethods();
         Mockito.verify(this.typesInGeneralConfigPart).withSubtypeResolver(Mockito.any());
@@ -74,13 +78,24 @@ public class JacksonModuleTest {
         Mockito.verifyNoMoreInteractions(this.configBuilder, this.fieldConfigPart, this.methodConfigPart, this.typesInGeneralConfigPart);
     }
 
-
     @Test
-    public void testApplyToConfigBuilderWithSkipSubtypeLookupAndIgnoreTypeInfoTranformOptions() {
-        new JacksonModule(JacksonOption.SKIP_SUBTYPE_LOOKUP, JacksonOption.IGNORE_TYPE_INFO_TRANSFORM)
+    public void testApplyToConfigBuilderWithRespectJsonPropertyOrderOption() {
+        new JacksonModule(JacksonOption.RESPECT_JSONPROPERTY_ORDER, JacksonOption.SKIP_SUBTYPE_LOOKUP, JacksonOption.IGNORE_TYPE_INFO_TRANSFORM)
                 .applyToConfigBuilder(this.configBuilder);
 
-        this.verifyCommonConfigurations();
+        this.verifyCommonConfigurations(true);
+
+        Mockito.verify(this.typesInGeneralConfigPart).withPropertySorter(Mockito.any(JsonPropertySorter.class));
+
+        Mockito.verifyNoMoreInteractions(this.configBuilder, this.fieldConfigPart, this.methodConfigPart, this.typesInGeneralConfigPart);
+    }
+
+    @Test
+    public void testApplyToConfigBuilderWithoutOptionalFeatures() {
+        new JacksonModule(JacksonOption.IGNORE_PROPERTY_NAMING_STRATEGY, JacksonOption.SKIP_SUBTYPE_LOOKUP, JacksonOption.IGNORE_TYPE_INFO_TRANSFORM)
+                .applyToConfigBuilder(this.configBuilder);
+
+        this.verifyCommonConfigurations(false);
 
         Mockito.verifyNoMoreInteractions(this.configBuilder, this.fieldConfigPart, this.methodConfigPart, this.typesInGeneralConfigPart);
     }
@@ -90,7 +105,7 @@ public class JacksonModuleTest {
         new JacksonModule(JacksonOption.SKIP_SUBTYPE_LOOKUP)
                 .applyToConfigBuilder(this.configBuilder);
 
-        this.verifyCommonConfigurations();
+        this.verifyCommonConfigurations(true);
 
         Mockito.verify(this.configBuilder).forMethods();
         Mockito.verify(this.typesInGeneralConfigPart).withCustomDefinitionProvider(Mockito.any());
@@ -105,7 +120,7 @@ public class JacksonModuleTest {
         new JacksonModule(JacksonOption.IGNORE_TYPE_INFO_TRANSFORM)
                 .applyToConfigBuilder(this.configBuilder);
 
-        this.verifyCommonConfigurations();
+        this.verifyCommonConfigurations(true);
 
         Mockito.verify(this.configBuilder).forMethods();
         Mockito.verify(this.typesInGeneralConfigPart).withSubtypeResolver(Mockito.any());
@@ -129,7 +144,7 @@ public class JacksonModuleTest {
         new JacksonModule(options)
                 .applyToConfigBuilder(this.configBuilder);
 
-        this.verifyCommonConfigurations();
+        this.verifyCommonConfigurations(true);
 
         Mockito.verify(this.configBuilder).forMethods();
         Mockito.verify(this.typesInGeneralConfigPart).withSubtypeResolver(Mockito.any());
@@ -142,40 +157,43 @@ public class JacksonModuleTest {
         Mockito.verifyNoMoreInteractions(this.configBuilder, this.fieldConfigPart, this.methodConfigPart, this.typesInGeneralConfigPart);
     }
 
-    private void verifyCommonConfigurations() {
+    private void verifyCommonConfigurations(boolean considerNamingStrategy) {
         Mockito.verify(this.configBuilder).getObjectMapper();
         Mockito.verify(this.configBuilder).forFields();
         Mockito.verify(this.configBuilder).forTypesInGeneral();
 
         Mockito.verify(this.fieldConfigPart).withDescriptionResolver(Mockito.any());
         Mockito.verify(this.fieldConfigPart).withIgnoreCheck(Mockito.any());
-        Mockito.verify(this.fieldConfigPart).withPropertyNameOverrideResolver(Mockito.any());
+        Mockito.verify(this.fieldConfigPart, Mockito.times(considerNamingStrategy ? 2 : 1)).withPropertyNameOverrideResolver(Mockito.any());
 
         Mockito.verify(this.typesInGeneralConfigPart).withDescriptionResolver(Mockito.any());
     }
 
     Object parametersForTestPropertyNameOverride() {
         return new Object[][]{
-            {"unannotatedField", null},
-            {"fieldWithEmptyPropertyAnnotation", null},
-            {"fieldWithSameValuePropertyAnnotation", null},
-            {"fieldWithNameOverride", "field override 1"},
-            {"fieldWithNameOverrideOnGetter", "method override 1"},
-            {"fieldWithNameOverrideAndOnGetter", "field override 2"}
+            {"unannotatedField", null, "unannotated-field"},
+            {"fieldWithEmptyPropertyAnnotation", null, "field-with-empty-property-annotation"},
+            {"fieldWithSameValuePropertyAnnotation", null, "field-with-same-value-property-annotation"},
+            {"fieldWithNameOverride", "field override 1", "field-with-name-override"},
+            {"fieldWithNameOverrideOnGetter", "method override 1", "field-with-name-override-on-getter"},
+            {"fieldWithNameOverrideAndOnGetter", "field override 2", "field-with-name-override-and-on-getter"}
         };
     }
 
     @Test
     @Parameters
-    public void testPropertyNameOverride(String fieldName, String expectedOverrideValue) throws Exception {
+    public void testPropertyNameOverride(String fieldName, String expectedOverrideValue, String kebabCaseName) throws Exception {
         new JacksonModule().applyToConfigBuilder(this.configBuilder);
 
         ArgumentCaptor<ConfigFunction<FieldScope, String>> captor = ArgumentCaptor.forClass(ConfigFunction.class);
-        Mockito.verify(this.fieldConfigPart).withPropertyNameOverrideResolver(captor.capture());
+        Mockito.verify(this.fieldConfigPart, Mockito.times(2)).withPropertyNameOverrideResolver(captor.capture());
 
         FieldScope field = new TestType(TestClassForPropertyNameOverride.class).getMemberField(fieldName);
-        String overrideValue = captor.getValue().apply(field);
-        Assert.assertEquals(expectedOverrideValue, overrideValue);
+        List<String> overrideValues = captor.getAllValues().stream()
+                .map(nameOverride -> nameOverride.apply(field))
+                .collect(Collectors.toList());
+        Assert.assertEquals(expectedOverrideValue, overrideValues.get(0));
+        Assert.assertEquals(kebabCaseName, overrideValues.get(1));
     }
 
     Object parametersForTestDescriptionResolver() {
@@ -224,6 +242,7 @@ public class JacksonModuleTest {
         Assert.assertEquals(expectedDescription, description);
     }
 
+    @JsonNaming(PropertyNamingStrategy.KebabCaseStrategy.class)
     private static class TestClassForPropertyNameOverride {
 
         Integer unannotatedField;